Put script_catch_errors in utils.py. Move uses into the scripts/ dirs. Adjust tests.
