+++ submit_cgap/scripts/make_sample_fastq_file.py
@@ -1,6 +1,7 @@
 import argparse
 
 from dcicutils.data_utils import generate_sample_fastq_file
+from ..utils import script_catch_errors
 
 
 EPILOG = __doc__
@@ -16,7 +17,9 @@ def main(simulated_args_for_testing=None):
     parser.add_argument('--number', '-n', help='number of sequences', default=10, type=int)
     parser.add_argument('--length', '-l', help='length of sequences', default=10, type=int)
     args = parser.parse_args(args=simulated_args_for_testing)
-    generate_sample_fastq_file(filename=args.filename, num=args.number, length=args.length)
+
+    with script_catch_errors():
+        generate_sample_fastq_file(filename=args.filename, num=args.number, length=args.length)
 
 
 if __name__ == '__main__':
