+++ submit_cgap/scripts/submit_metadata_bundle.py
@@ -1,6 +1,7 @@
 import argparse
 from ..submission import submit_any_ingestion, DEFAULT_INGESTION_TYPE
 from ..base import UsingCGAPKeysFile
+from ..utils import script_catch_errors
 
 
 EPILOG = __doc__
@@ -28,12 +29,14 @@ def main(simulated_args_for_testing=None):
                         help="search subfolders of folder for upload files", default=False)
     args = parser.parse_args(args=simulated_args_for_testing)
 
-    return submit_any_ingestion(ingestion_filename=args.bundle_filename, ingestion_type=args.ingestion_type,
-                                institution=args.institution, project=args.project,
-                                server=args.server, env=args.env,
-                                validate_only=args.validate_only, upload_folder=args.upload_folder,
-                                no_query=args.no_query, subfolders=args.subfolders
-                                )
+    with script_catch_errors():
+
+        submit_any_ingestion(ingestion_filename=args.bundle_filename, ingestion_type=args.ingestion_type,
+                             institution=args.institution, project=args.project,
+                             server=args.server, env=args.env,
+                             validate_only=args.validate_only, upload_folder=args.upload_folder,
+                             no_query=args.no_query, subfolders=args.subfolders,
+                             )
 
 
 if __name__ == '__main__':
