+++ submit_cgap/tests/test_make_sample_fastq_file.py
@@ -1,6 +1,7 @@
 from unittest import mock
 from ..scripts.make_sample_fastq_file import main as make_sample_fastq_file_main
 from ..scripts import make_sample_fastq_file as make_sample_fastq_file_module
+from .testing_helpers import system_exit_expected
 
 
 def test_make_sample_fastq_file_script():
@@ -8,11 +9,9 @@ def test_make_sample_fastq_file_script():
     def test_it(args_in, expect_exit_code, expect_called, expect_call_args=None):
         with mock.patch.object(make_sample_fastq_file_module,
                                "generate_sample_fastq_file") as mock_generate_sample_fastq_file:
-            try:
+            with system_exit_expected(exit_code=expect_exit_code):
                 make_sample_fastq_file_main(args_in)
                 mock_generate_sample_fastq_file.assert_called_with(**expect_call_args)
-            except SystemExit as e:
-                assert e.code == expect_exit_code
             assert mock_generate_sample_fastq_file.call_count == (1 if expect_called else 0)
 
     test_it(args_in=[], expect_exit_code=2, expect_called=False)  # Missing args
